Radiology/requirements.txt
h5py==2.10.0
matplotlib==3.3.4
omegaconf==2.0.6
pandas==1.1.4
pyaml==20.4.0
scikit-learn==0.24.2
scipy==1.5.4
tensorflow==1.14
tqdm==4.60.0